refactor: extract solana.js from main.js (BAT-201)
Move Solana RPC, base58 encoding, transaction building, Jupiter DEX
(token resolution, quotes, swaps, prices, triggers, recurring),
wallet management into solana.js module.

- Self-contained module with no circular deps
- Depends on: config.js, web.js (httpRequest), bridge.js (androidBridgeCall)
- 24 exports covering all Solana/Jupiter functionality
- main.js: 6054 → 5278 lines (-776)
